app/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.1
async-timeout==4.0.2
asyncio==3.4.3
attrs==21.4.0
beautifulsoup4==4.11.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==3.2
django-debug-toolbar==3.2.4
django-extensions==3.1.5
django-split-settings==1.1.0
flake8==4.0.0
frozenlist==1.3.0
gunicorn==20.1.0
idna==3.3
iniconfig==1.1.1
lxml==4.8.0
mccabe==0.6.1
multidict==6.0.2
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
pyparsing==3.0.8
pytest==7.0.1
python-dotenv==0.19.2
pytz==2022.1
requests==2.27.1
soupsieve==2.3.2.post1
sqlparse==0.4.2
tomli==2.0.1
urllib3==1.26.9
yarl==1.7.2
